backend/requirements.txt
# AI Healthcare Assistant - Backend Requirements
# Voice interaction dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Voice Processing
openai==1.3.7  # For Whisper API
elevenlabs==0.2.27  # For Text-to-Speech
pydub==0.25.1  # Audio processing
python-multipart==0.0.6  # File uploads

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Original dependencies
chatterbot==0.8.4
SQLAlchemy
gunicorn==22.0.0